back/pyproject.toml
[project]
name = "flask-poetry-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Anderson Silva",email = "andersoonn.ams@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"

dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-restful (>=0.3.10,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "annotated-types (==0.6.0)",
    "flask-pymongo (>=3.0.1,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.0.0"
pytest-flask = "^1.3.0"
pytest-mock = "^3.12.0"

[tool.pytest.ini_options]
testpaths = ["tests", "app/api/tasks/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.poetry.scripts]
test = "pytest:main"
